fix(dns): isolate UDP DNS resolution
