+++ crates/honk-tool/src/sub.rs
@@ -5,8 +5,10 @@
 //! and IPv6 (a full protocol dial through the node), and a proxied latency
 //! measurement (`urltest_node`).
 
+mod udp;
+
 use std::io::Read as _;
-use std::net::{IpAddr, SocketAddr};
+use std::net::SocketAddr;
 use std::sync::Arc;
 use std::time::{Duration, Instant};
 
@@ -23,6 +25,10 @@ use honk_outbound::reality::parse_reality_config;
 use honk_outbound::urltest::urltest_node;
 use url::Url;
 use uuid::Uuid;
+
+use udp::{
+    UdpCheckTarget, parse_udp_check_target, probe_udp_dns, probe_udp_quic, system_dns_resolver,
+};
 
 #[derive(Clone, Copy, Debug, Eq, PartialEq, ValueEnum)]
 pub enum TlsImplementation {
@@ -641,581 +647,5 @@ fn split_host_port(s: &str) -> anyhow::Result<(&str, u16)> {
     Ok((host, port.parse()?))
 }
 
-#[derive(Debug, Eq, PartialEq)]
-enum UdpCheckTarget {
-    Literal(SocketAddr),
-    Host { host: String, port: u16 },
-}
-
-fn parse_udp_check_target(targets: &[String]) -> anyhow::Result<UdpCheckTarget> {
-    for target in targets {
-        let target = target.trim();
-        if let Ok(address) = target.parse::<SocketAddr>() {
-            return Ok(UdpCheckTarget::Literal(address));
-        }
-        if let Ok(ip) = target.parse::<IpAddr>() {
-            return Ok(UdpCheckTarget::Literal(SocketAddr::new(ip, 53)));
-        }
-    }
-    let target = targets
-        .first()
-        .map(String::as_str)
-        .context("--udp-check requires at least one target")?
-        .trim();
-    let endpoint = honk_core::dns::endpoint::DnsEndpoint::parse(
-        target,
-        honk_config::types::DnsProtocol::Udp,
-        None,
-    )?;
-    Ok(UdpCheckTarget::Host {
-        host: endpoint.host,
-        port: endpoint.port,
-    })
-}
-
-// Avoid uncancellable getaddrinfo work and public-resolver fallback.
-fn system_dns_resolver() -> Option<DnsResolver> {
-    let contents = std::fs::read_to_string("/etc/resolv.conf").ok()?;
-    let server = contents.lines().find_map(|line| {
-        let mut fields = line.split_whitespace();
-        if fields.next()? != "nameserver" {
-            return None;
-        }
-        fields
-            .next()?
-            .parse::<IpAddr>()
-            .ok()
-            .map(|ip| SocketAddr::new(ip, 53))
-    })?;
-    let mut config = honk_config::dns::DnsConfig {
-        upstream: vec![honk_config::dns::DnsUpstream {
-            name: "system".into(),
-            address: server.to_string(),
-            protocol: honk_config::types::DnsProtocol::Udp,
-            tls_server_name: None,
-            outbound: None,
-        }],
-        ..Default::default()
-    };
-    config.routing.fallback = "system".into();
-    if std::path::Path::new(honk_config::dns::SYSTEM_HOSTS_PATH).is_file() {
-        config
-            .hosts
-            .push(honk_config::dns::SYSTEM_HOSTS_PATH.into());
-    }
-    DnsResolver::new(&config).ok()
-}
-
-async fn resolve_udp_check_target(
-    target: &UdpCheckTarget,
-    resolver: Option<&DnsResolver>,
-) -> Result<SocketAddr, ProbeFailureKind> {
-    match target {
-        UdpCheckTarget::Literal(address) => Ok(*address),
-        UdpCheckTarget::Host { host, port } => {
-            let Some(resolver) = resolver else {
-                return Err(ProbeFailureKind::Resolve);
-            };
-            let resolved = resolver
-                .resolve(host)
-                .await
-                .map_err(|_| ProbeFailureKind::Resolve)?;
-            resolved
-                .ipv4
-                .into_iter()
-                .chain(resolved.ipv6)
-                .next()
-                .map(|ip| SocketAddr::new(ip, *port))
-                .ok_or(ProbeFailureKind::Resolve)
-        }
-    }
-}
-
-/// Tiny xorshift PRNG seeded from the clock (avoids a rand dependency for the
-/// two probe packet builders).
-fn next_rand(state: &mut u64) -> u64 {
-    let mut x = *state;
-    x ^= x << 13;
-    x ^= x >> 7;
-    x ^= x << 17;
-    *state = x;
-    x
-}
-
-fn rand_seed() -> u64 {
-    std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map(|d| d.as_nanos() as u64 | 1)
-        .unwrap_or(0x9e3779b97f4a7c15)
-}
-
-fn build_dns_probe_query(id: u16) -> Vec<u8> {
-    let mut query = vec![
-        (id >> 8) as u8,
-        id as u8,
-        0x01,
-        0x00,
-        0x00,
-        0x01,
-        0x00,
-        0x00,
-        0x00,
-        0x00,
-        0x00,
-        0x00,
-    ];
-    for label in ["google", "com"] {
-        query.push(label.len() as u8);
-        query.extend_from_slice(label.as_bytes());
-    }
-    query.extend_from_slice(&[0x00, 0x00, 0x01, 0x00, 0x01]);
-    query
-}
-
-/// UDP probe: one minimal DNS A query through the node's UDP transport.
-/// Proves the node's UDP relay path end to end (mirrors the engine's
-/// `probe_node_udp` health check).
-async fn probe_udp_dns(
-    registry: &ProxyRegistry,
-    node: &Node,
-    target: &UdpCheckTarget,
-    resolver: Option<&DnsResolver>,
-    timeout: Duration,
-) -> Option<Result<Duration, ProbeFailureKind>> {
-    let Some(entry) = registry.find(node.protocol()) else {
-        return Some(Err(ProbeFailureKind::Handler));
-    };
-    if !(entry.descriptor.supports_udp)(node) {
-        return None;
-    }
-    let packet = entry.packet.as_ref()?;
-    let result = tokio::time::timeout(timeout, async {
-        let dns_server = resolve_udp_check_target(target, resolver).await?;
-        let transport = packet
-            .dial_udp_transport(node, dns_server, None, timeout)
-            .await
-            .map_err(|_| ProbeFailureKind::Exchange)?;
-
-        let mut rng = rand_seed();
-        let id = next_rand(&mut rng) as u16;
-        let query = build_dns_probe_query(id);
-
-        let start = Instant::now();
-        transport
-            .send_packet(&query)
-            .await
-            .map_err(|_| ProbeFailureKind::Exchange)?;
-        let mut buf = [0u8; 512];
-        match transport.recv_packet(&mut buf).await {
-            Ok((n, _)) if n >= 2 && buf[0] == query[0] && buf[1] == query[1] => Ok(start.elapsed()),
-            Ok(_) | Err(_) => Err(ProbeFailureKind::Exchange),
-        }
-    })
-    .await;
-    Some(match result {
-        Ok(result) => result,
-        Err(_) => Err(ProbeFailureKind::Timeout),
-    })
-}
-
-/// UDP probe for QUIC: run a real QUIC handshake through the node's UDP
-/// transport and time it.  Unlike a bare Version-Negotiation trigger (which
-/// most frontends silently drop), this proves TLS-in-QUIC reachability
-/// through the node's UDP path.
-async fn probe_udp_quic(
-    registry: &ProxyRegistry,
-    node: &Node,
-    url_host: &str,
-    url_port: u16,
-    timeout: Duration,
-) -> Option<Result<Duration, ProbeFailureKind>> {
-    let Some(entry) = registry.find(node.protocol()) else {
-        return Some(Err(ProbeFailureKind::Handler));
-    };
-    if !(entry.descriptor.supports_udp)(node) {
-        return None;
-    }
-    let packet = entry.packet.as_ref()?;
-    let addr = match tokio::net::lookup_host((url_host, url_port)).await {
-        Ok(mut addrs) => addrs.find(SocketAddr::is_ipv4)?,
-        Err(_) => return Some(Err(ProbeFailureKind::Resolve)),
-    };
-    let transport = match packet.dial_udp_transport(node, addr, None, timeout).await {
-        Ok(transport) => transport,
-        Err(_) => return Some(Err(ProbeFailureKind::Exchange)),
-    };
-
-    let probe_node = Node {
-        outbound: honk_config::node::OutboundConfig::Hysteria2(
-            honk_config::node::Hysteria2Config {
-                quic: honk_config::node::QuicOptions {
-                    tls: honk_config::node::TlsOptions {
-                        skip_cert_verify: true,
-                        sni: Some(url_host.to_string()),
-                        ..Default::default()
-                    },
-                    ..Default::default()
-                },
-                ..Default::default()
-            },
-        ),
-        ..Default::default()
-    };
-    let config = match honk_outbound::quic::client_config(
-        &probe_node,
-        &[b"h3"],
-        honk_outbound::quic::QuicClientOptions::default(),
-    )
-    .await
-    {
-        Ok(config) => config,
-        Err(_) => return Some(Err(ProbeFailureKind::Exchange)),
-    };
-
-    Some(
-        honk_outbound::quic::quic_handshake_probe(transport, addr, url_host, &config, timeout)
-            .await
-            .map_err(|_| ProbeFailureKind::Exchange),
-    )
-}
-
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn udp_check_target_prefers_literals_and_defaults_to_dns_port() {
-        for (targets, expected) in [
-            (vec!["1.2.3.4"], "1.2.3.4:53"),
-            (vec!["2001:db8::1"], "[2001:db8::1]:53"),
-            (vec!["[2001:db8::1]:5353"], "[2001:db8::1]:5353"),
-            (vec!["my.dns.invalid", "9.9.9.9"], "9.9.9.9:53"),
-        ] {
-            let targets = targets.into_iter().map(str::to_string).collect::<Vec<_>>();
-            assert_eq!(
-                parse_udp_check_target(&targets).unwrap(),
-                UdpCheckTarget::Literal(expected.parse().unwrap())
-            );
-        }
-        assert_eq!(
-            parse_udp_check_target(&["my.dns.invalid:5353".to_string()]).unwrap(),
-            UdpCheckTarget::Host {
-                host: "my.dns.invalid".into(),
-                port: 5353,
-            }
-        );
-        assert!(parse_udp_check_target(&[]).is_err());
-        assert!(parse_udp_check_target(&[String::new()]).is_err());
-    }
-
-    fn silent_dns_resolver(address: SocketAddr) -> DnsResolver {
-        let mut config = honk_config::dns::DnsConfig {
-            upstream: vec![honk_config::dns::DnsUpstream {
-                name: "test".into(),
-                address: address.to_string(),
-                protocol: honk_config::types::DnsProtocol::Udp,
-                tls_server_name: None,
-                outbound: None,
-            }],
-            ..Default::default()
-        };
-        config.routing.fallback = "test".into();
-        DnsResolver::new(&config).unwrap()
-    }
-
-    fn direct_node() -> Node {
-        let mut node = Node {
-            name: "direct-test".into(),
-            host: "127.0.0.1".into(),
-            outbound: honk_config::node::OutboundConfig::from_protocol(NodeProtocol::Direct),
-            ..Default::default()
-        };
-        node.id = node.derive_id();
-        node
-    }
-
-    #[tokio::test]
-    async fn udp_dns_resolution_timeout_does_not_abort_unrelated_probe() {
-        let sink = tokio::net::UdpSocket::bind("127.0.0.1:0").await.unwrap();
-        let closed_port = tokio::net::TcpSocket::new_v4().unwrap();
-        closed_port.bind("127.0.0.1:0".parse().unwrap()).unwrap();
-        let address = closed_port.local_addr().unwrap();
-        let targets = ProbeTargets {
-            host: address.ip().to_string(),
-            port: address.port(),
-            url: Some(format!("http://{address}/")),
-            timeout: Duration::from_millis(40),
-            v4: Some(address),
-            v6: Some(address),
-            udp_dns: UdpCheckTarget::Host {
-                host: "dns.test".into(),
-                port: 53,
-            },
-            dns_resolver: Some(Arc::new(silent_dns_resolver(sink.local_addr().unwrap()))),
-        };
-        let registry = ProxyRegistry::default_resolver().unwrap();
-        let outcome = probe_node(&registry, direct_node(), &targets).await;
-        assert_eq!(outcome.udp_dns, Some(Err(ProbeFailureKind::Timeout)));
-        assert_eq!(outcome.urltest, Some(Err(ProbeFailureKind::Exchange)));
-        assert_eq!(outcome.v4, Some(Err(ProbeFailureKind::Exchange)));
-    }
-
-    #[tokio::test]
-    async fn udp_dns_without_system_resolver_reports_resolve() {
-        let registry = ProxyRegistry::default_resolver().unwrap();
-        let node = direct_node();
-        let target = UdpCheckTarget::Host {
-            host: "dns.test".into(),
-            port: 53,
-        };
-        assert_eq!(
-            probe_udp_dns(&registry, &node, &target, None, Duration::from_millis(40),).await,
-            Some(Err(ProbeFailureKind::Resolve))
-        );
-    }
-
-    #[tokio::test]
-    async fn udp_ineligible_node_skips_dns_target_resolution() {
-        let sink = tokio::net::UdpSocket::bind("127.0.0.1:0").await.unwrap();
-        let resolver = silent_dns_resolver(sink.local_addr().unwrap());
-        let target = UdpCheckTarget::Host {
-            host: "dns.test".into(),
-            port: 53,
-        };
-        let registry = ProxyRegistry::default_resolver().unwrap();
-        assert_eq!(
-            probe_udp_dns(
-                &registry,
-                &vless_node(),
-                &target,
-                Some(&resolver),
-                Duration::from_millis(40),
-            )
-            .await,
-            None
-        );
-        let mut packet = [0u8; 512];
-        assert!(
-            tokio::time::timeout(Duration::from_millis(20), sink.recv(&mut packet))
-                .await
-                .is_err()
-        );
-    }
-
-    #[test]
-    fn dns_probe_query_has_full_header_and_google_a_question() {
-        let query = build_dns_probe_query(0x1234);
-        assert_eq!(&query[..12], &[0x12, 0x34, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0]);
-        assert_eq!(&query[12..], b"\x06google\x03com\x00\x00\x01\x00\x01");
-    }
-
-    fn vless_node() -> Node {
-        Node {
-            name: "vless-test".into(),
-            address: "192.0.2.1:443".into(),
-            host: "192.0.2.1".into(),
-            port: 443,
-            outbound: honk_config::node::OutboundConfig::Vless(honk_config::node::VlessConfig {
-                uuid: Some("b831381d-6324-4d53-ad4f-8cda48b30811".into()),
-                tls: honk_config::node::TlsOptions {
-                    enabled: true,
-                    ..Default::default()
-                },
-                ..Default::default()
-            }),
-            ..Default::default()
-        }
-    }
-
-    #[test]
-    fn registry_contains_node_dependent_vless_udp() {
-        let registry = ProxyRegistry::default_resolver().unwrap();
-        let entry = registry.find(NodeProtocol::VLess).unwrap();
-        assert_eq!(entry.descriptor.protocol, NodeProtocol::VLess);
-        assert!(entry.probeable.is_some());
-        assert!(entry.packet.is_some());
-        let mut node = vless_node();
-        assert!(!(entry.descriptor.supports_udp)(&node));
-        node.vless_mut().unwrap().mode = WireMode::UotV2;
-        assert!((entry.descriptor.supports_udp)(&node));
-    }
-
-    #[test]
-    fn stdin_subscription_url_rules() {
-        assert_eq!(
-            parse_subscription_url_from_stdin("  https://example.com/feed?token=value\n").unwrap(),
-            "https://example.com/feed?token=value"
-        );
-        assert_eq!(
-            parse_subscription_url_from_stdin("http://127.0.0.1/sub").unwrap(),
-            "http://127.0.0.1/sub"
-        );
-        for invalid in [
-            "",
-            "   \n",
-            "not a URL",
-            "ftp://example.com/sub",
-            "https:///",
-            "https://one.example/sub\nhttps://two.example/sub",
-        ] {
-            assert_eq!(
-                parse_subscription_url_from_stdin(invalid)
-                    .unwrap_err()
-                    .to_string(),
-                "invalid subscription URL from stdin"
-            );
-        }
-    }
-
-    #[test]
-    fn vless_probe_eligibility_precedence_and_reasons() {
-        assert_eq!(
-            classify_vless_node(&vless_node()),
-            ProbeEligibility::Supported
-        );
-
-        let mut node = vless_node();
-        let vless = node.vless_mut().unwrap();
-        vless.uuid = None;
-        vless.tls.reality_short_id = Some("abc".into());
-        assert_eq!(
-            classify_vless_node(&node),
-            ProbeEligibility::InvalidConfig("invalid-uuid")
-        );
-
-        let mut node = vless_node();
-        let vless = node.vless_mut().unwrap();
-        vless.tls.reality_short_id = Some("abc".into());
-        vless.transport.transport = "kcp".into();
-        assert_eq!(
-            classify_vless_node(&node),
-            ProbeEligibility::InvalidConfig("invalid-reality")
-        );
-
-        let mut node = vless_node();
-        node.vless_mut().unwrap().transport.transport = "kcp".into();
-        assert_eq!(
-            classify_vless_node(&node),
-            ProbeEligibility::ExpectedUnsupported("unsupported-transport")
-        );
-
-        let mut node = vless_node();
-        node.vless_mut().unwrap().flow = Some("unsupported-flow-value".into());
-        assert_eq!(
-            classify_vless_node(&node),
-            ProbeEligibility::ExpectedUnsupported("unsupported-flow")
-        );
-
-        let mut node = vless_node();
-        let vless = node.vless_mut().unwrap();
-        vless.tls.enabled = false;
-        vless.flow = Some("xtls-rprx-vision".into());
-        assert_eq!(
-            classify_vless_node(&node),
-            ProbeEligibility::InvalidConfig("vision-without-tls")
-        );
-
-        let mut node = vless_node();
-        let vless = node.vless_mut().unwrap();
-        vless.transport.transport = "ws".into();
-        vless.flow = Some("xtls-rprx-vision".into());
-        assert_eq!(
-            classify_vless_node(&node),
-            ProbeEligibility::ExpectedUnsupported("vision-non-tcp")
-        );
-
-        let mut node = vless_node();
-        node.name.clear();
-        assert_eq!(
-            classify_vless_node(&node),
-            ProbeEligibility::InvalidConfig("invalid-config")
-        );
-    }
-
-    #[test]
-    fn vless_shapes_are_fixed_and_non_identifying() {
-        let mut node = vless_node();
-        let vless = node.vless_mut().unwrap();
-        vless.tls.enabled = false;
-        vless.transport.transport.clear();
-        assert_eq!(vless_shape(&node), "vless/plain/tcp");
-
-        node.vless_mut().unwrap().mode = WireMode::Xudp;
-        assert_eq!(vless_shape(&node), "vless/plain/tcp/xudp");
-        node.vless_mut().unwrap().mode = WireMode::MuxCool;
-        assert_eq!(vless_shape(&node), "vless/plain/tcp/mux-cool");
-        node.vless_mut().unwrap().mode = WireMode::Legacy;
-
-        let vless = node.vless_mut().unwrap();
-        vless.tls.enabled = true;
-        vless.transport.transport = "ws".into();
-        assert_eq!(vless_shape(&node), "vless/tls/ws");
-
-        let vless = node.vless_mut().unwrap();
-        vless.tls.reality_public_key = Some("private-key-material".into());
-        vless.transport.transport = "grpc".into();
-        vless.flow = Some("xtls-rprx-vision".into());
-        assert_eq!(vless_shape(&node), "vless/reality/grpc/vision");
-
-        let vless = node.vless_mut().unwrap();
-        vless.transport.transport = "kcp-secret".into();
-        vless.flow = Some("provider-flow-secret".into());
-        assert_eq!(vless_shape(&node), "vless/reality/unsupported");
-    }
-
-    #[test]
-    fn vless_udp_is_rendered_as_not_applicable() {
-        let registry = ProxyRegistry::default_resolver().unwrap();
-        let outcome = ProbeOutcome::timed_out(&registry, &vless_node());
-        assert_eq!(outcome.udp_dns, None);
-        assert_eq!(outcome.udp_quic, None);
-        let rendered = render_outcome(&outcome);
-        assert!(rendered.contains("dns: n/a"));
-        assert!(rendered.contains("quic: n/a"));
-    }
-
-    #[test]
-    fn vless_udp_mode_is_rendered_as_probeable() {
-        let registry = ProxyRegistry::default_resolver().unwrap();
-        let mut node = vless_node();
-        node.vless_mut().unwrap().mode = WireMode::H2mux;
-        let outcome = ProbeOutcome::timed_out(&registry, &node);
-        assert!(matches!(
-            outcome.udp_dns,
-            Some(Err(ProbeFailureKind::Timeout))
-        ));
-        assert!(render_outcome(&outcome).contains("dns: FAIL(timeout)"));
-        assert_eq!(outcome.shape, "vless/tls/tcp/h2mux");
-    }
-
-    #[test]
-    fn rendered_failures_exclude_connection_identifiers() {
-        let mut node = vless_node();
-        node.host = "sentinel-host.invalid".into();
-        node.address = "sentinel-host.invalid:443".into();
-        let vless = node.vless_mut().unwrap();
-        vless.uuid = Some("sentinel-uuid".into());
-        vless.tls.sni = Some("sentinel-sni.invalid".into());
-        vless.tls.reality_public_key = Some("sentinel-reality-key".into());
-        let outcome = ProbeOutcome::skipped(&node, classify_vless_node(&node));
-        let rendered = render_outcome(&outcome);
-        for sentinel in [
-            "https://sentinel-url.invalid/private?token=secret",
-            "sentinel-host.invalid",
-            "sentinel-uuid",
-            "sentinel-sni.invalid",
-            "sentinel-reality-key",
-        ] {
-            assert!(!rendered.contains(sentinel));
-        }
-
-        for (kind, expected) in [
-            (ProbeFailureKind::Resolve, "FAIL(resolve)"),
-            (ProbeFailureKind::Timeout, "FAIL(timeout)"),
-            (ProbeFailureKind::Exchange, "FAIL(exchange)"),
-            (ProbeFailureKind::Handler, "FAIL(handler)"),
-        ] {
-            assert_eq!(render_probe_result(&Some(Err(kind)), "n/a"), expected);
-        }
-    }
-}
+mod tests;
